+++ projects/ng-helpers/src/directives/debounce-change/debounce-change.directive.ts
@@ -1,20 +1,17 @@
-import {AfterViewInit, Directive, ElementRef, EventEmitter, Inject, Input, NgZone, Output} from '@angular/core';
-import {fromEvent} from 'rxjs';
-import {debounceTime, filter, takeUntil} from 'rxjs/operators';
-import {RxDestroy} from '../../helpers/rx-destroy';
+import {AfterViewInit, Directive, ElementRef, EventEmitter, Inject, Input, NgZone, OnDestroy, Output} from '@angular/core';
+import {fromEvent, Subscription} from 'rxjs';
+import {debounceTime, filter} from 'rxjs/operators';
 import {DEBOUNCE_TIME} from './debounce-time.const';
 
 @Directive({
   selector: '[jpDebounceChange]'
 })
-export class DebounceChangeDirective extends RxDestroy implements AfterViewInit {
+export class DebounceChangeDirective implements AfterViewInit, OnDestroy {
   constructor(
     private _el: ElementRef,
     private _ngZone: NgZone,
     @Inject(DEBOUNCE_TIME) private _defaultDebounceTime: number
-  ) {
-    super();
-  }
+  ) {}
 
   /**
    * time to forward to the debounceTime pipe
@@ -37,14 +34,15 @@ export class DebounceChangeDirective extends RxDestroy implements AfterViewInit 
    */
   @Output() jpDebounceChange = new EventEmitter<string>();
 
+  subscription: Subscription;
+
   ngAfterViewInit() {
     this._ngZone.runOutsideAngular(() => {
 
       let prev = this._el.nativeElement.value;
 
-      fromEvent<any>(this._el.nativeElement, this.debounceChangeEventType)
+      this.subscription = fromEvent<any>(this._el.nativeElement, this.debounceChangeEventType)
         .pipe(
-          takeUntil(this.destroyed$),
           debounceTime(this.debounceTime || this._defaultDebounceTime),
           filter(event => {
             return event.target &&
@@ -64,5 +62,11 @@ export class DebounceChangeDirective extends RxDestroy implements AfterViewInit 
           });
         });
     });
+  }
+
+  ngOnDestroy() {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
   }
 }