+++ projects/ng-helpers/src/directives/load-click/load-click.directive.ts
@@ -1,7 +1,6 @@
-import {Directive, ElementRef, Inject, Input, OnInit, Renderer2} from '@angular/core';
-import {Observable} from 'rxjs';
-import {finalize, takeUntil} from 'rxjs/operators';
-import {RxDestroy} from '../../helpers/rx-destroy';
+import {Directive, ElementRef, Inject, Input, OnDestroy, OnInit, Renderer2} from '@angular/core';
+import {Observable, Subscription} from 'rxjs';
+import {finalize} from 'rxjs/operators';
 import {LOAD_CLICK_CLASS} from './load-click-class.const';
 
 /**
@@ -11,14 +10,12 @@ import {LOAD_CLICK_CLASS} from './load-click-class.const';
  */
 
 @Directive({selector: '[jpLoadClick]'})
-export class LoadClickDirective extends RxDestroy implements OnInit {
+export class LoadClickDirective implements OnInit, OnDestroy {
   constructor(
     private _el: ElementRef,
     private _renderer: Renderer2,
     @Inject(LOAD_CLICK_CLASS) private _defaultLoadClickClass: string
-  ) {
-    super();
-  }
+  ) {}
 
   @Input()
   jpLoadClick: () => Observable<any>;
@@ -37,6 +34,8 @@ export class LoadClickDirective extends RxDestroy implements OnInit {
 
   @Input()
   disableAttribute = true;
+
+  subscription: Subscription;
 
   ngOnInit() {
     this._renderer.listen(
@@ -63,7 +62,7 @@ export class LoadClickDirective extends RxDestroy implements OnInit {
           );
         }
 
-        this.jpLoadClick()
+        this.subscription = this.jpLoadClick()
           .pipe(
             finalize(() => {
               this._renderer.removeClass(this._el.nativeElement, defaultClass);
@@ -71,11 +70,16 @@ export class LoadClickDirective extends RxDestroy implements OnInit {
               if (this.disableAttribute) {
                 this._renderer.removeAttribute(this._el.nativeElement, 'disabled');
               }
-            }),
-            takeUntil(this.destroyed$)
+            })
           )
           .subscribe();
       }
     );
+  }
+
+  ngOnDestroy() {
+    if (this.subscription) {
+      this.subscription.unsubscribe();
+    }
   }
 }